Changed packages structure. Moved Book related DTOs to dto.book package.
