+++ src/main/java/mate/academy/onlinebookstore/dto/book/BookDto.java
@@ -1,4 +1,4 @@
-package mate.academy.onlinebookstore.dto;
+package mate.academy.onlinebookstore.dto.book;
 
 import java.math.BigDecimal;
 import lombok.Data;
