+++ src/main/java/mate/academy/onlinebookstore/dto/book/BookSearchParametersDto.java
@@ -1,4 +1,4 @@
-package mate.academy.onlinebookstore.dto;
+package mate.academy.onlinebookstore.dto.book;
 
 public record BookSearchParametersDto(String[] titles, String[] authors, String[] isbns) {
 }
