+++ src/main/java/mate/academy/onlinebookstore/dto/book/CreateBookRequestDto.java
@@ -1,4 +1,4 @@
-package mate.academy.onlinebookstore.dto;
+package mate.academy.onlinebookstore.dto.book;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
