+++ src/main/java/mate/academy/onlinebookstore/mapper/BookMapper.java
@@ -1,8 +1,8 @@
 package mate.academy.onlinebookstore.mapper;
 
 import mate.academy.onlinebookstore.config.MapperConfig;
-import mate.academy.onlinebookstore.dto.BookDto;
-import mate.academy.onlinebookstore.dto.CreateBookRequestDto;
+import mate.academy.onlinebookstore.dto.book.BookDto;
+import mate.academy.onlinebookstore.dto.book.CreateBookRequestDto;
 import mate.academy.onlinebookstore.model.Book;
 import org.mapstruct.Mapper;
