+++ src/main/java/mate/academy/onlinebookstore/repository/SpecificationBuilder.java
@@ -1,6 +1,6 @@
 package mate.academy.onlinebookstore.repository;
 
-import mate.academy.onlinebookstore.dto.BookSearchParametersDto;
+import mate.academy.onlinebookstore.dto.book.BookSearchParametersDto;
 import org.springframework.data.jpa.domain.Specification;
 
 public interface SpecificationBuilder<T> {
